Adjust to moving of test classes to different package
